fix!: `pip-extra-index-url` no longer needs a quote at the end
In the action.yml file, a `"` character was missing. This forced
users to add one to the end of the input variable which is
confusing since it does not look like proper yaml on the surface.
Unfortunately, this is a breaking change, because anyone updating
will need to remove the quote that they added.